+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1] - 2022-06-06
+
+### Fixed
+
+- Filter error if no quotes have been created yet
+
+## [1.1.0] - 2022-05-19
+
+### Added
+
+- Improved organization/cost center filtering on `QuotesTable` component
+
 ## [1.0.1] - 2022-04-26
 
 ### Fixed